CMN-213: move files after rebase with release
