+++ src/stories/01 Atoms/button-favourite/button-favourite.scss
