+++ src/stories/03 Components/search-result-item/SearchResultItem.stories.tsx
