+++ src/stories/03 Components/search-result-item/search-result-item.scss
